requirements.txt
# === ML LIBRARIES (Modernized) ===
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
imbalanced-learn==0.11.0

# === DATA VALIDATION ===
pandera==0.17.2

# === EXPERIMENT TRACKING & REGISTRY (Modernized) ===
mlflow==2.13.0
dvc==3.36.1
dvc-s3==3.0.1

# === API & DEPLOYMENT (Cleaned) ===
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.3
pydantic-settings==2.1.0

# === ORCHESTRATION (Modernized) ===
prefect==2.14.20

# === MONITORING (Modernized) ===
evidently>=0.4.19
prometheus-client==0.19.0

# === TESTING & QUALITY ===
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# === UTILITIES ===
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
loguru==0.7.2
